MAJOR rewrite of http backend
HTTP and WS backends and frontend are now in one repository.
Both backends were doing similar jobs and now are easier to maintain.

HTTP backend is fairly modular and can be extended through different
handlers. It is also based on templates, so web pages are now dynamic.

Authorization and authentication is based on cookies, and are fairly
secure. For example, cookies do not leak any information about other
database contents, e.g. user count or even which users token it is.